Tetris v4.0 - Initial Release
Everything is now up and running correctly. Training takes a while but the program does learn to clear lines over time. Program now also has saving/storing functionality so that it can resume training on different machines from the same spot. Training can also be temporarily paused by pressing esc. By default, the parameters are saved every 10,000 time steps but this is easily adjustable on line 687. This will be the "initial" release of the software for public use.