Tetris v4.1 - Prioritized Experience Replay and Q-Value Logging
Added Prioritized Experience Replay so the agent will now always train off of experiences where it cleared a line if any are present in the replay memory, otherwise it will use random sampling. The batch size will still always be the same size and if there are not enough lines cleared to fill up the prioritized experience replay it will be padded with random samples. In addition, there are now options to enable logging of each actions Q-Values. The reward value was also adjusted slightly to penalize higher pieces more.