See extended description before use. Use Tetris.py for now.
This is the same as Tetris v4.2 but it trains the network based on the experiences with the largest error between their estimated rewards and their actual rewards. This works but the network is over trained from this and as a result will clear 0 lines. This is more of a proof of concept although I may mess with it later on to get it to a working point.